ROFO-215 refactor: FoodSpotDetailReview -> ExpendableReview 이름 변경
