+++ presentation/src/main/java/com/weit2nd/presentation/model/reivew/ExpendableReview.kt
@@ -1,8 +1,8 @@
-package com.weit2nd.presentation.ui.foodspot.detail
+package com.weit2nd.presentation.model.reivew
 
 import com.weit2nd.presentation.model.foodspot.Review
 
-data class FoodSpotDetailReview(
+data class ExpendableReview(
     val review: Review,
     val isExpended: Boolean,
 )
